+++ src/Pages/Setup.jsx
@@ -11,22 +11,9 @@ import BoxSetup from '../Components/BoxSetup'
 import CastingSetup from '../Components/CastingSetup'
 import Slide from 'react-reveal/Slide';
 import { ButtonBase } from '@mui/material';
-import MediaQuery from 'react-responsive'
-import { Select, FormControl, InputLabel, MenuItem } from '@mui/material';
-import { ThemeProvider, createTheme } from '@mui/material/styles';
-import CssBaseline from '@mui/material/CssBaseline';
 
-const darkTheme = createTheme({
-    palette: {
-        mode: 'dark',
-    },
-});
 
-const lightTheme = createTheme({
-    palette: {
-        mode: 'light',
-    },
-});
+
 
 const Setup = ({ darkMode, setDarkMode }) => {
     const [current, setCurrent] = useState('Web');
@@ -42,11 +29,7 @@ const Setup = ({ darkMode, setDarkMode }) => {
     const BoxRef = useRef(null);
     const CastingRef = useRef(null);
 
-    const [muiSelectValue, setMuiSelectValue] = React.useState('Web');
-    const handleChange = (e) => {
-        setMuiSelectValue(e.target.value)
-        setCurrent(e.target.value)
-    }
+
 
     return (
         
@@ -55,7 +38,7 @@ const Setup = ({ darkMode, setDarkMode }) => {
             <h1 className='text-center font-bold text-5xl  mt-20  text-white'>DOWNLOAD THE BEST IPTV APP</h1>
             <p className='text-center font-bold text-2xl mt-4  text-white'>Our Apps Work Flawlessly on tv, mobile and browser</p>
 
-            <MediaQuery minWidth={980}>
+            
                 <div className='mt-24  w-full md:w-[90%]  lg:max-w-[1100px] flex-wrap rounded-mg shadow-sm mx-auto flex items-center justify-between flex-col gap-8'>
                     <div className='w-full h-full p-2 justify-between  flex-wrap md:flex-row flex-col flex  gap-8'>
 
@@ -163,56 +146,9 @@ const Setup = ({ darkMode, setDarkMode }) => {
                     </div>
                 </div>
 
-            </MediaQuery>
+           
             
-            <MediaQuery minWidth={0} maxWidth={979}>
-
-            {darkMode?  
-                
-                <ThemeProvider theme={darkTheme}>
-                    <CssBaseline />
-                    <FormControl className='mt-24  w-full md:w-[90%]  lg:max-w-[1100px] '>
-                        <InputLabel id="demo-simple-select-label">Screen</InputLabel>
-                        <Select
-                            labelId="demo-simple-select-label"
-                            id="demo-simple-select"
-                            value={muiSelectValue}
-                            label="Screen"
-                            onChange={handleChange}
-                        >
-                            <MenuItem value={'Web'}>Web</MenuItem>
-                            <MenuItem value={'Mobile'}>Mobile</MenuItem>
-                            <MenuItem value={'TV'}>TV</MenuItem>
-                            <MenuItem value={'Box & Stick'}>Box & Stick</MenuItem>
-                            <MenuItem value={'Casting'}>Casting</MenuItem>
-                        </Select>
-                    </FormControl>
-                </ThemeProvider>
-           :
             
-            <ThemeProvider theme={lightTheme}>
-                    <CssBaseline />
-                    <FormControl className='mt-24  w-full md:w-[90%]  lg:max-w-[1100px] '>
-                        <InputLabel id="demo-simple-select-label">Screen</InputLabel>
-                        <Select
-                            labelId="demo-simple-select-label"
-                            id="demo-simple-select"
-                            value={muiSelectValue}
-                            label="Screen"
-                            onChange={handleChange}
-                        >
-                            <MenuItem value={'Web'}>Web</MenuItem>
-                            <MenuItem value={'Mobile'}>Mobile</MenuItem>
-                            <MenuItem value={'TV'}>TV</MenuItem>
-                            <MenuItem value={'Box & Stick'}>Box & Stick</MenuItem>
-                            <MenuItem value={'Casting'}>Casting</MenuItem>
-                        </Select>
-                    </FormControl>
-                </ThemeProvider>
-         
-            
-            }
-              </MediaQuery>
        
             <div className='mt-6 max-w-full md:w-[90%]  lg:max-w-[1100px] dark:bg-blue-darkmd bg-white dark:text-white transition-all duration-300 ease-in-out text-blue-lightmd rounded-lg shadow-lg mx-auto flex items-center justify-center p-8'>
                 {current === 'Web' && (prevIndex >= index)? <Slide duration= {700} enter = 'false' left> <div ref={WebRef}><WebSetup /></div> </Slide > : ''}
